% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMAToolsTesting-package.R
\docType{package}
\name{BMAToolsTesting-package}
\alias{BMAToolsTesting}
\alias{BMAToolsTesting-package}
\title{BMAToolsTesting: Testing BMA Specifications}
\description{
Testing BMA Specifications.
}
\author{
\strong{Maintainer}: Matthew Clarke \email{matthewaclarke1991@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
