% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_funcs.r
\name{check_conflicts}
\alias{check_conflicts}
\title{Check for conflicts between perturbed nodes and background, ensures
that backgrounds take precedence.}
\usage{
check_conflicts(results, backgrounds, node_col = "name", drugs = NULL)
}
\arguments{
\item{results}{Results from \code{\link{process_results}}}

\item{backgrounds}{Backgrounds}

\item{node_col}{node column name in backgrounds}

\item{drugs}{optional drugs data frame}
}
\description{
Check for conflicts between perturbed nodes and background, ensures
that backgrounds take precedence.
}
