% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_funcs.r
\name{check_drug_nodes}
\alias{check_drug_nodes}
\title{Check whether drugs applied to the network have corresponding nodes
in the network}
\usage{
check_drug_nodes(drugs, netw_variables, node_col_name)
}
\arguments{
\item{drugs}{data frame of drugs to apply with columns: 'drug' with
the name of the drug; 'node' with the node(s) affect (for
multiple nodes specify one node per row with the same drug
name) and 'activity' the numerical level to which the drug
should set the affected node}

\item{netw_variables}{data frame of node names, ids, range_from and
range_to for a network, generated by
\code{\link{get_netw_variables}}}

\item{node_col_name}{name of the columns specifying nodes}
}
\value{
Stop if nodes are missing from network, and report missing nodes
}
\description{
Check whether drugs applied to the network have corresponding nodes
in the network
}
\seealso{
Other combination_therapy: 
\code{\link{check_drug_conflicts}()},
\code{\link{check_drugs_in_range}()},
\code{\link{get_drugs_commands}()},
\code{\link{make_bkg_commands_combo}()},
\code{\link{make_command_args}()},
\code{\link{make_pair_muts}()},
\code{\link{make_single_muts}()},
\code{\link{process_results}()},
\code{\link{run_all_backgrounds}()}
}
\concept{combination_therapy}
