% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_funcs.r
\name{check_drugs_in_range}
\alias{check_drugs_in_range}
\title{Check whether drugs applied to the network have effects that are
withing the range of the node. If not, BioCheckConsole would fail
silently.}
\usage{
check_drugs_in_range(drugs_commands)
}
\arguments{
\item{drugs_commands}{drugs with network variables and command
fragments, generated by \code{\link{get_netw_variables}}}
}
\value{
Stop if there are perturbations out of range and report
}
\description{
Check whether drugs applied to the network have effects that are
withing the range of the node. If not, BioCheckConsole would fail
silently.
}
\seealso{
Other combination_therapy: 
\code{\link{check_drug_conflicts}()},
\code{\link{check_drug_nodes}()},
\code{\link{get_drugs_commands}()},
\code{\link{make_bkg_commands_combo}()},
\code{\link{make_command_args}()},
\code{\link{make_pair_muts}()},
\code{\link{make_single_muts}()},
\code{\link{process_results}()},
\code{\link{run_all_backgrounds}()}
}
\concept{combination_therapy}
