% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_perts_in_range.r
\name{check_perts_in_range}
\alias{check_perts_in_range}
\title{check_perts_in_range}
\usage{
check_perts_in_range(spec_levels)
}
\arguments{
\item{spec_levels}{specification dataframe generated by
\code{\link{convert_spec_levels}}}
}
\value{
Stop if perturbations are outside range of nodes
}
\description{
Check that all perturbations are within the range of the node
affected. Otherwise BMA will run without perturbation silently.
}
