% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spec_groups.r
\name{check_spec_groups}
\alias{check_spec_groups}
\title{check_spec_groups}
\usage{
check_spec_groups(spec, group_vars)
}
\arguments{
\item{spec}{specification loaded through \code{\link{import_spec}}}

\item{group_vars}{columns of specification used to group experiments}
}
\value{
Stop if specification grouping likely to misgroup experiments
}
\description{
Test whether separate experiments have the same grouping factor,
and so would be run together in spec testing and possibly give
misleading results
}
\details{
I expect all rows of an experiment to be together. Therefore the
group index for each row should appear all in a group, and at most
once. So first I remove all consecutive duplicates, but not
non-consecutive duplicates i.e. 3 3 3 4 4 4 3 3 3 becomes 3 4
3. If there is a non-consecutive duplication, this suggests two
seperate experiments share a group key, and the spec should be
ammended either a) to put all experimental specification for a
single experiment on consecutive rows, or b to change the
experiment_particular to make clear the difference between the two
experiments. So I remove consecutive dupes, find the remaining
dupes, match this to the group keys to give a descriptive error
message of where the duplications are, and throw and error to
force this to be fixed.
}
