% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_funcs.r
\name{combo}
\alias{combo}
\title{combo}
\usage{
combo(
  netw_file_path,
  backgrounds_path,
  drug_path,
  bma_path = "C:\\\\\\"Program Files (x86)\\"\\\\BMA\\\\BioCheckConsole.exe",
  results_prefix = "COMBO_RUN",
  out_dir,
  project_path = "",
  node_col_name = "node",
  use_vmcai = TRUE,
  pheno_only = FALSE,
  phenotypes = NA,
  use_exclusions = FALSE,
  exclusions_path = NA,
  drug_conflict_overide = FALSE,
  skip_drugs_single = FALSE,
  skip_drugs_pairs = FALSE,
  skip_all_pairs = FALSE,
  short_filenames = FALSE,
  log_filename
)
}
\arguments{
\item{netw_file_path}{path to network JSON file}

\item{backgrounds_path}{path to backgrounds (CSV)}

\item{drug_path}{path to list of drug perturbations to be applied
in addition to node perturbations (CSV)}

\item{bma_path}{path to BMA command line installation, defaults to
the path produced by the one click installer (.msi). The path
is automatically normalized for cross-platform compatibility.}

\item{results_prefix}{prefix to results directory}

\item{out_dir}{output directory}

\item{project_path}{project path for git SHA log, point to git
repo of the network and specification being tested}

\item{node_col_name}{name of the node column in the mutational
background file and drug file}

\item{use_vmcai}{Use VMCAI mode}

\item{pheno_only}{Only evaluate the level of phenotype nodes
supplied by the phenotypes argument. Speeds up processing step
and reduces size of final processed_data file}

\item{phenotypes}{list of phenotypes in format \code{c("node_name_1", "node_name_2")} for use if pheno_only is TRUE}

\item{use_exclusions}{exclude perturbation of a set of nodes
supplied to the \code{exclusions_path} argument. Speeds up
processing by skipping those nodes which are uninteresting to
perturb e.g. phenotype nodes}

\item{exclusions_path}{list of nodes in format \code{c("node_name_1", "node_name_2")} for use if use_exclusions is TRUE}

\item{drug_conflict_overide}{allow overide of drug conflicts
check, only used for testing}

\item{skip_drugs_single}{skip drug perturbations}

\item{skip_drugs_pairs}{skip drug perturbations}

\item{skip_all_pairs}{skip pairwise combinations of nodes}

\item{short_filenames}{logical. If TRUE, use MD5 hashes for long filenames
to avoid Windows path length issues. Defaults to FALSE.}

\item{log_filename}{filename for log file}
}
\value{
write a \code{parsed_results.csv} and \code{processed_results.csv}
}
\description{
Run all node and drug combinations (separately) for a given
network on a set of mutational backgrounds
}
