% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_funcs.r
\name{combo_parallel}
\alias{combo_parallel}
\title{Run combination screen in parallel across different backgrounds}
\usage{
combo_parallel(
  netw_file_path,
  combo_backgrounds_path,
  n_cores = parallel::detectCores() - 1,
  results_prefix = "COMBO_RUN",
  out_dir,
  project_path = "",
  combo_drug_path = NA,
  bma_path = "C:\\\\\\"Program Files (x86)\\"\\\\BMA\\\\BioCheckConsole.exe",
  node_col_name = "node",
  use_vmcai = TRUE,
  pheno_only = TRUE,
  phenotypes = c("output_a", "output_b"),
  use_exclusions = FALSE,
  exclusions_path = NA,
  drug_conflict_overide = TRUE,
  skip_all_pairs = FALSE,
  skip_combo_drugs_single = TRUE,
  skip_combo_drugs_double = TRUE,
  log_filename = "Combo.log",
  short_filenames = FALSE,
  ...
)
}
\arguments{
\item{netw_file_path}{path to network JSON file}

\item{combo_backgrounds_path}{path to backgrounds CSV file containing
multiple backgrounds}

\item{n_cores}{number of cores to use for parallel processing, defaults
to detectCores() - 1}

\item{results_prefix}{prefix for results directory name}

\item{out_dir}{output directory where results will be stored}

\item{project_path}{project path for git SHA log, point to git
repo of the network and specification being tested}

\item{combo_drug_path}{path to list of drug perturbations to be applied
in addition to node perturbations (CSV), defaults to NA}

\item{bma_path}{path to BMA command line installation, defaults to
the path produced by the one click installer (.msi)}

\item{node_col_name}{name of the node column in the mutational
background file and drug file, defaults to "node"}

\item{use_vmcai}{Use VMCAI mode, defaults to TRUE}

\item{pheno_only}{Only evaluate the level of phenotype nodes
supplied by the phenotypes argument, defaults to TRUE}

\item{phenotypes}{list of phenotypes in format c("node_name_1",
"node_name_2") for use if pheno_only is TRUE}

\item{use_exclusions}{exclude perturbation of a set of nodes
supplied to the exclusions_path argument, defaults to FALSE}

\item{exclusions_path}{path to CSV file with nodes to exclude}

\item{drug_conflict_overide}{allow overide of drug conflicts
check, only used for testing, defaults to TRUE}

\item{skip_all_pairs}{skip pairwise combinations of nodes,
defaults to FALSE}

\item{skip_combo_drugs_single}{skip single drug perturbations,
defaults to TRUE}

\item{skip_combo_drugs_double}{skip double drug perturbations,
defaults to TRUE}

\item{log_filename}{filename for log file, defaults to "Combo.log"}

\item{short_filenames}{logical. If TRUE, use MD5 hashes for long filenames
to avoid Windows path length issues. Defaults to FALSE.}

\item{...}{Additional arguments (for backwards compatibility).
The deprecated arguments 'git_log' and 'bma_tools_path' are accepted
but ignored with a warning. Other arguments will cause an error.}
}
\value{
Creates results directories for each background and writes
integrated results files: parsed_integrated_results.csv,
node_integrated_results.csv, and processed_integrated_results.csv
}
\description{
This function parallelizes the combo workflow by splitting backgrounds
and running each background in a separate parallel worker. It handles
temporary file creation, parallel execution, post-processing, and
results integration.
}
\seealso{
Other combination_therapy:
\code{\link[=check_drug_conflicts]{check_drug_conflicts()}},
\code{\link[=check_drug_nodes]{check_drug_nodes()}},
\code{\link[=check_drugs_in_range]{check_drugs_in_range()}},
\code{\link[=get_drugs_commands]{get_drugs_commands()}},
\code{\link[=make_bkg_commands_combo]{make_bkg_commands_combo()}},
\code{\link[=make_command_args]{make_command_args()}},
\code{\link[=make_pair_muts]{make_pair_muts()}},
\code{\link[=make_single_muts]{make_single_muts()}},
\code{\link[=process_results]{process_results()}},
\code{\link[=run_all_backgrounds]{run_all_backgrounds()}}
}
\concept{combination_therapy}
