% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_spec_levels.r
\name{convert_spec_levels}
\alias{convert_spec_levels}
\title{convert_spec_levels}
\usage{
convert_spec_levels(spec, log_file)
}
\arguments{
\item{spec}{specification dataframe generated by
\code{\link{import_spec}}}

\item{log_file}{name of log file}
}
\value{
specification data frame with absolute numerical values
for perturbation and expectation to be passed to
BioCheckConsole
}
\description{
Allow the conversion of a specification written with relative
terms (min, mid, max) for the perturbation and expectation to be
converted to absolute numbers for passing to BioCheckConsole
(BMA). This is safer as it will continue to produce the expected
result even when the granularity of a node changes.
}
