% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest_filename.r
\name{digest_filename}
\alias{digest_filename}
\title{digest_filename}
\usage{
digest_filename(filename, append_json = FALSE)
}
\arguments{
\item{filename}{filename}

\item{append_json}{Add ".json" after digesting}
}
\value{
shortened filename
}
\description{
Use MD5 hash to create a unique short filename for long
perturbations to avoid long paths causing issues on Windows
}
