% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_hashtables.r
\name{get_all_hashtables}
\alias{get_all_hashtables}
\title{get_all_hashtables}
\usage{
get_all_hashtables(file_hashtable_dir, log_file)
}
\arguments{
\item{file_hashtable_dir}{directory for hashtables}

\item{log_file}{log filename}
}
\value{
all hashes in a single table
}
\description{
get_all_hashtables
}
