% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_funcs.r
\name{get_combo_results_dir}
\alias{get_combo_results_dir}
\title{get_combo_results_dir}
\usage{
get_combo_results_dir(results_prefix, project_path, out_dir, netw_file_path)
}
\arguments{
\item{results_prefix}{prefix to results directory}

\item{project_path}{project path for git SHA log, point to git
repo of the network and specification being tested}

\item{out_dir}{directory where all output files should be stored}

\item{netw_file_path}{path to network JSON file}
}
\value{
string with path to results directory
}
\description{
Generate results directory for combinations in a standardised way,
so that visualisation scripts can generate one that lines up with
\code{\link{combo}} without combo having to pass it on
}
