% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_funcs.r
\name{get_drugs_commands}
\alias{get_drugs_commands}
\title{Get the BioCheckConsole command line commands to effect the drug
perturbations, and assign filename parts to record perturbations
applied}
\usage{
get_drugs_commands(drugs, netw_variables, node_col_name)
}
\arguments{
\item{drugs}{data frame of drugs to apply with columns: 'drug' with
the name of the drug; 'node' with the node(s) affect (for
multiple nodes specify one node per row with the same drug
name) and 'activity' the numerical level to which the drug
should set the affected node. Drug names will be sanitised with
janitor::make_clean_names as they must be passed to the command
line}

\item{netw_variables}{data frame of node names, ids, range_from and
range_to for a network, generated by
\code{\link{get_netw_variables}}}

\item{node_col_name}{name of the columns specifying nodes}
}
\value{
Data frame of: 'drug' sanitised drug name;
'drug_name_original' original name, 'node' node affected,
'activity' activity of the node under the effects of the drug,
'id' BMA node id, 'range_from' lower bound of range of node in
network, 'range_to' upper bound of range of node in network,
'formula' network target function, 'command arg' command to
cause perturbation in format '-ko id activity',
'alt_filename_part' part of the filename for the results file
corresponding to this perturbation
}
\description{
Get the BioCheckConsole command line commands to effect the drug
perturbations, and assign filename parts to record perturbations
applied
}
\seealso{
Other combination_therapy: 
\code{\link{check_drug_conflicts}()},
\code{\link{check_drug_nodes}()},
\code{\link{check_drugs_in_range}()},
\code{\link{combo_parallel}()},
\code{\link{make_bkg_commands_combo}()},
\code{\link{make_command_args}()},
\code{\link{make_pair_muts}()},
\code{\link{make_single_muts}()},
\code{\link{process_results}()},
\code{\link{run_all_backgrounds}()}
}
\concept{combination_therapy}
