% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netw_variables.R
\name{get_netw_variables}
\alias{get_netw_variables}
\title{Get BMA network variables}
\usage{
get_netw_variables(netw_file_path)
}
\arguments{
\item{netw_file_path}{File path to a BMA JSON file}
}
\value{
A tibble of the name (name), lower bound of the range (range_from),
upper bound of the range (range_to) and Target Function (formula)
}
\description{
Imports the node names, target functions and ranges from a BMA json
as a tibble
}
