% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_funcs.r
\name{import_drugs_clean}
\alias{import_drugs_clean}
\title{import_drugs_clean}
\usage{
import_drugs_clean(drug_path, show_col_types = FALSE)
}
\arguments{
\item{drug_path}{path to drugs CSV}

\item{show_col_types}{Mutes printing of columns types using readr::read_csv}
}
\value{
tibble of drugs with drug names sanitised
}
\description{
Import drugs and sanitise drug names so they cannot cause problems
when converted to file names
}
