% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_spec.r
\name{import_spec}
\alias{import_spec}
\title{import spec}
\usage{
import_spec(spec_path, loserum, clean_underscores, netw_variables)
}
\arguments{
\item{spec_path}{path to specification csv file}

\item{loserum}{option to set all serum nodes to 1 (experimental)}

\item{clean_underscores}{remove all underscores at start of a
string in source and cell line columns (as often conversion to
csv adds them)}

\item{netw_variables}{network variables dataframe generated by
\code{\link{get_netw_variables}}}
}
\value{
specification data frame
}
\description{
import specification for autopert testing
}
