% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.r
\name{loop_plot_diff_double}
\alias{loop_plot_diff_double}
\title{loop_plot_diff_double}
\usage{
loop_plot_diff_double(
  bkg,
  df,
  phenos,
  levs,
  neg_col,
  pos_col,
  cluster_rows,
  cluster_cols,
  results_dir,
  vis_dir,
  suffix,
  h = 25,
  w = 25,
  fontsize = 10,
  drug_type = "",
  node_col = "node"
)
}
\arguments{
\item{bkg}{background to plot}

\item{df}{dataframe of processed combination results}

\item{phenos}{phenotypes to visualise. Selects from \code{node} column of
processed results}

\item{levs}{level of the mutations to visualise e.g. 0,
c(1:max_level) or "" for drugs}

\item{neg_col}{negative colour}

\item{pos_col}{positive colour}

\item{cluster_rows}{cluster_rows TRUE/FALSE}

\item{cluster_cols}{cluster columns (TRUE/FALSE)}

\item{results_dir}{results directory}

\item{vis_dir}{subdirectory for heatmaps}

\item{suffix}{suffix for filenames}

\item{h}{height of heatmap}

\item{w}{width of heatmap}

\item{fontsize}{fontsize of heatmap}

\item{drug_type}{encoded in filename}

\item{node_col}{node column name in imported data}
}
\value{
write heatmaps to file
}
\description{
Loop over plotting of double diff heatmaps
}
