% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_funcs.r
\name{make_bkg_commands_combo}
\alias{make_bkg_commands_combo}
\title{Convert a list of background mutations for combination screen}
\usage{
make_bkg_commands_combo(backgrounds, netw_variables, node_col = "name")
}
\arguments{
\item{backgrounds}{data frame as above}

\item{netw_variables}{data from of network variables generated by
\code{\link{get_netw_variables}}}

\item{node_col}{name of the column with the node
names. \code{\link{get_netw_variables}} defaults to \code{name}.}
}
\value{
data frame with annotation of all the arguments to pass to
BioCheckConsole for each background, and a filename part that
describes these. Unlike \code{\link{make_command_args}} these
are grouped so that each background is a single row, even if
this is multiple perturbations of the network.
}
\description{
Convert a 3 x n dataframe of backgrounds in long format, where one
column is the name of the background called "background", one is a
list of nodes called "name" and one is the activity the nodes
should be fixed to in this background called "activity"
}
\seealso{
Other combination_therapy: 
\code{\link{check_drug_conflicts}()},
\code{\link{check_drug_nodes}()},
\code{\link{check_drugs_in_range}()},
\code{\link{combo_parallel}()},
\code{\link{get_drugs_commands}()},
\code{\link{make_command_args}()},
\code{\link{make_pair_muts}()},
\code{\link{make_single_muts}()},
\code{\link{process_results}()},
\code{\link{run_all_backgrounds}()}
}
\concept{combination_therapy}
