% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_funcs.r
\name{make_command_args}
\alias{make_command_args}
\title{Take a df of nodes from json file, annotated with level of activity
and annotate all with filename prefix and BioCheckConsole command
argument needed to change it's value}
\usage{
make_command_args(
  df,
  id_col = "id",
  activity_col = "activity",
  node_col = "name"
)
}
\arguments{
\item{df}{dataframe of nodes and ids generated by
\code{\link{get_netw_variables}}}

\item{id_col}{name of column in \code{df} containing BMA node ids, as a
string, default "id"}

\item{activity_col}{name of column in \code{df} containing activity
level to change node to in perturbations, as a string, default
"activity"}

\item{node_col}{name of column in \code{df} containing the name of the
node/gene from BMA network, as a string, default "name"}
}
\value{
dataframe with a column of arguments to pass to a call of
BioCheckConsole, \code{command_arg} and a column of the filename
prefixes for the output of this command, \code{filename_part}
}
\description{
Take a df of nodes from json file, annotated with level of activity
and annotate all with filename prefix and BioCheckConsole command
argument needed to change it's value
}
\seealso{
\code{\link{get_netw_variables}}

Other combination_therapy: 
\code{\link{check_drug_conflicts}()},
\code{\link{check_drug_nodes}()},
\code{\link{check_drugs_in_range}()},
\code{\link{combo_parallel}()},
\code{\link{get_drugs_commands}()},
\code{\link{make_bkg_commands_combo}()},
\code{\link{make_pair_muts}()},
\code{\link{make_single_muts}()},
\code{\link{process_results}()},
\code{\link{run_all_backgrounds}()}
}
\concept{combination_therapy}
