% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_funcs.r
\name{make_pair_drugs}
\alias{make_pair_drugs}
\title{Make full commands for pair drug perturbations by combining all
perturbations for a pair drug}
\usage{
make_pair_drugs(drugs_single)
}
\arguments{
\item{drugs_single}{drugs with network variables and command
fragments, generated by \code{\link{make_single_drugs}}}
}
\value{
Dataframe of drug commands for use by
\code{\link{run_all_backgrounds}}
}
\description{
Make full commands for pair drug perturbations by combining all
perturbations for a pair drug
}
