% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_funcs.r
\name{make_single_drugs}
\alias{make_single_drugs}
\title{Make full commands for single drug perturbations by combining all
perturbations for a single drug}
\usage{
make_single_drugs(drugs_commands)
}
\arguments{
\item{drugs_commands}{drugs with network variables and command
fragments, generated by \code{\link{get_netw_variables}}}
}
\value{
Dataframe of drug commands for use by
\code{\link{run_all_backgrounds}}
}
\description{
Make full commands for single drug perturbations by combining all
perturbations for a single drug
}
