% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_funcs.r
\name{make_single_muts}
\alias{make_single_muts}
\title{Generate a dataframe with columns describing for BMA all possible
single activations and inhibitions of nodes in a network}
\usage{
make_single_muts(netw_variables, node_col = "name")
}
\arguments{
\item{netw_variables}{data frame of node names, ids, range_from and
range_to for a network, generated by
\code{\link{get_netw_variables}}}

\item{node_col}{name of the column with the node
names. \code{\link{get_netw_variables}} defaults to \code{name}.}
}
\value{
data frame annotated by command arguments and filenames for
perturbations of all nodes
}
\description{
Generate a dataframe with columns describing for BMA all possible
single activations and inhibitions of nodes in a network
}
\seealso{
\code{\link{make_command_args}},
\code{\link{get_netw_variables}}

Other combination_therapy: 
\code{\link{check_drug_conflicts}()},
\code{\link{check_drug_nodes}()},
\code{\link{check_drugs_in_range}()},
\code{\link{get_drugs_commands}()},
\code{\link{make_bkg_commands_combo}()},
\code{\link{make_command_args}()},
\code{\link{make_pair_muts}()},
\code{\link{process_results}()},
\code{\link{run_all_backgrounds}()}
}
\concept{combination_therapy}
