% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NANSEN-package.R
\docType{package}
\name{NANSEN-package}
\alias{NANSEN}
\alias{NANSEN-package}
\title{NANSEN: NANSEN (Network Analysis aNd ScrEeNing)}
\description{
A package for verification and analysis of biological gene regulatory network models built with Bio Model Analyzer.
}
\keyword{internal}
