% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_name_check.R
\name{node_name_check}
\alias{node_name_check}
\title{Check BMA node names}
\usage{
node_name_check(name)
}
\arguments{
\item{name}{name/id of a BMA node}
}
\value{
None
}
\description{
Checks for disallowed characters in nodes used in an imported BMA network
}
