% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_biocheck.R
\name{parse_biocheck_dir}
\alias{parse_biocheck_dir}
\title{parse_biocheck_dir}
\usage{
parse_biocheck_dir(dir, netw_variables, rec = FALSE)
}
\arguments{
\item{dir}{path to a directory containing BioCheck JSON results files}

\item{netw_variables}{dataframe of BMA JSON variables, created with
\code{get_netw_variables}}

\item{rec}{boolean to specify whether to apply recursively. If TRUE, stores
directory structure in "filename" column of output.}
}
\value{
dataframe of:

"filename" filename of the json from which the results come (path from dir to
file if \code{rec = TRUE})

"time" time steps taken to restrict values as much as possible;

"id" id number of node;

"lo" lower bound to which node value was restricted;

"hi" upper bound to which node value was restricted
"name" node name

"range_from" lower bound of unrestricted node range

"range_to" upper bound of unrestricted node range

"formula" node target function
}
\description{
Get attractors from BioCheck VMCAI results json files in directory
}
\details{
Finds the values to which the range of all nodes can be restricted by the
VMCAI algorithm of BMA BioCheck, at the last timepoint outputted, and stores
as one dataframe, using the filename as a reference column
Also makes more human readable by getting the relevant information
per node from the network_variables.
Ignores cex files.
}
