% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_biocheck.R
\name{parse_biocheck_dir_apend}
\alias{parse_biocheck_dir_apend}
\title{parse_biocheck_dir_apend}
\usage{
parse_biocheck_dir_apend(existing_file, dir, netw_variables, rec = FALSE)
}
\arguments{
\item{existing_file}{filename of existing file to apend to}

\item{dir}{directory of results JSON files}

\item{netw_variables}{network variables from \code{get_netw_variables}}

\item{rec}{apply recursively over a directory over directories}
}
\description{
parse_biocheck_dir_apend
}
