% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_biocheck.R
\name{parse_biocheck_json}
\alias{parse_biocheck_json}
\title{Get attractor from BioCheck VMCAI results json file}
\usage{
parse_biocheck_json(filepath)
}
\arguments{
\item{filepath}{path to a BioCheck JSON results file}
}
\value{
dataframe of:
"time" time steps taken to restrict values as much as possible;
"id" id number of node;
"lo" lower bound to which node value was restricted;
"hi" upper bound to which node value was restricted
}
\description{
Finds the values to which the range of all nodes can be restricted by the
VMCAI algorithm of BMA BioCheck, at the last timepoint outputted.
}
