% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.r
\name{plot_diff_double}
\alias{plot_diff_double}
\title{plot_diff_double}
\usage{
plot_diff_double(
  df,
  filename,
  results_dir,
  vis_dir,
  neg_col,
  pos_col,
  pheno,
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  levs,
  bkg,
  h = 25,
  w = 25,
  fontsize = 10,
  node_col = "node",
  mirror = TRUE,
  na_col = "black"
)
}
\arguments{
\item{df}{dataframe of processed combination results}

\item{filename}{filename to save heatmap}

\item{results_dir}{results directory}

\item{vis_dir}{subdirectory for heatmaps}

\item{neg_col}{negative colour}

\item{pos_col}{positive colour}

\item{pheno}{phenotype to visualise. Selects from \code{node} column of
processed results}

\item{cluster_rows}{cluster rows (TRUE/FALSE)}

\item{cluster_cols}{cluster columns (TRUE/FALSE)}

\item{levs}{level of the mutations to visualise e.g. 0,
c(1:max_level) or "" for drugs}

\item{bkg}{background to plot}

\item{h}{height of heatmap}

\item{w}{width of heatmap}

\item{fontsize}{fontsize of heatmap}

\item{node_col}{node column name in imported data}

\item{mirror}{Usually data is not mirrored in muta and mutb, and
so to plot a heatmap we do this within the function. For diff
plots it is useful to sometimes feed pre-mirrored data, so
this turns off this step}

\item{na_col}{set colour for na}
}
\value{
heatmap saved to drive, and a ggplot object for further
processing e.g. with patchwork
}
\description{
Plot a combination heatmap with a colourscheme optimised for
positive and negative data e.g. for relative levels rather than
absolute
}
