% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.r
\name{plot_diff_heatmaps}
\alias{plot_diff_heatmaps}
\title{plot_heatmaps}
\usage{
plot_diff_heatmaps(
  results_file,
  results_prefix,
  project_path,
  out_dir,
  netw_file_path,
  vis_dir,
  phenotypes,
  type = "node",
  neg_col = "Reds",
  pos_col = "Blues",
  cluster_rows,
  cluster_double_cols,
  background_order = NA,
  background_order_neat = NA,
  neaten_background = FALSE,
  single_fontsize = 10,
  single_druggable_fontsize = 25,
  single_drugs_fontsize = 20,
  double_fontsize = 10,
  double_druggable_font_size = 20,
  double_drugs_font_size = 20,
  h_s = 20,
  h_d = 25,
  w_s = 6,
  w_d = 25
)
}
\arguments{
\item{results_prefix}{prefix to results directory}

\item{project_path}{project path for git SHA log, point to git
repo of the network and specification being tested}

\item{out_dir}{output directory}

\item{netw_file_path}{path to network JSON file}

\item{vis_dir}{subdirectory for heatmaps}

\item{phenotypes}{list of phenotypes in format \code{c("node_name_1", "node_name_2")} for use if pheno_only is TRUE}

\item{type}{node, druggable or drug heatmaps}

\item{cluster_rows}{cluster_rows TRUE/FALSE}

\item{single_fontsize}{fontsize for monotherapy node perturbation heatmaps}

\item{single_druggable_fontsize}{fontsize for monotherapy druggable perturbation heatmaps}

\item{single_drugs_fontsize}{fontsize for monotherapy drug perturbation heatmaps}

\item{double_fontsize}{fontsize for combination therapy node perturbation heatmaps}

\item{double_druggable_font_size}{fontsize for combination therapy druggable perturbation heatmaps}

\item{double_drugs_font_size}{fontsize for combination therapy drug perturbation heatmaps}
}
\value{
plot heatmaps and save to directory
}
\description{
plot all heatmaps of a particular type, node, druggable or drug
}
