% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.r
\name{plot_heatmaps}
\alias{plot_heatmaps}
\title{plot_heatmaps}
\usage{
plot_heatmaps(
  results_file,
  results_prefix,
  project_path,
  out_dir,
  netw_file_path,
  vis_dir,
  cluster_rows = TRUE,
  cluster_double_cols = TRUE,
  type = "node",
  background_order = NA,
  background_order_neat = NA,
  neaten_background = FALSE,
  single_fontsize = 10,
  single_druggable_fontsize = 25,
  single_drugs_fontsize = 20,
  double_fontsize = 10,
  double_druggable_font_size = 20,
  double_drugs_font_size = 20,
  h_s = 20,
  h_d = 25,
  w_s = 6,
  w_d = 25
)
}
\arguments{
\item{results_prefix}{prefix to results directory}

\item{project_path}{project path for git SHA log, point to git
repo of the network and specification being tested}

\item{out_dir}{output directory}

\item{netw_file_path}{path to network JSON file}

\item{vis_dir}{subdirectory for heatmaps}

\item{cluster_rows}{cluster_rows TRUE/FALSE}
}
\value{
plot heatmaps and save to directory
}
\description{
plot all heatmaps of a particular type, node, druggable or drug
}
