% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.r
\name{plot_single}
\alias{plot_single}
\title{plot_single}
\usage{
plot_single(
  df,
  filename,
  results_dir,
  vis_dir,
  palette,
  cluster_rows = TRUE,
  pheno,
  levs,
  h = 20,
  w = 6,
  fontsize = 10,
  node_col = "node",
  na_col = "black"
)
}
\arguments{
\item{df}{dataframe of processed combination results}

\item{filename}{filename to save heatmap}

\item{results_dir}{results directory}

\item{vis_dir}{subdirectory for heatmaps}

\item{palette}{colour palette for heatmaps to be provided to
RColorBrewer::brewer.pal}

\item{cluster_rows}{cluster_rows TRUE/FALSE}

\item{pheno}{phenotype to visualise. Selects from \code{node} column of
processed results}

\item{levs}{level of the mutations to visualise e.g. 0,
c(1:max_level) or "" for drugs}

\item{h}{height of heatmap}

\item{w}{width of heatmap}

\item{fontsize}{fontsize of heatmap}

\item{node_col}{node column name in imported data}

\item{na_col}{set colour for na}
}
\value{
heatmap saved to drive, and a ggplot object for further
processing e.g. with patchwork
}
\description{
Plot monotherapy heatmap
}
