% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preserve_order.r
\name{preserve_order}
\alias{preserve_order}
\title{Set a column as a factor, which the levels in the current order}
\usage{
preserve_order(df, col)
}
\arguments{
\item{df}{dataframe in desired ordering}

\item{col}{column name which determines the order}
}
\value{
dataframe where col has been frozen in desired order as a factor
}
\description{
Use to preserve the ordering of a data frame, so that when it is passed
to a ggplot function the plotting preserves that order
}
