% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_funcs.r
\name{process_results}
\alias{process_results}
\title{Process the results of combination perturbation screen, assuming
two perturbations}
\usage{
process_results(
  parsed_results,
  phenotypes,
  log_file = log_file,
  node_col = "name",
  pheno_only = FALSE
)
}
\arguments{
\item{parsed_results}{raw results read from a directory of json
files using \code{\link{parse_biocheck_dir}}}

\item{phenotypes}{names of the nodes whose results will be analysed
e.g. "Proliferation"}

\item{log_file}{log file}

\item{node_col}{name of the column with the node
names. \code{\link{get_netw_variables}} defaults to \code{name}.}

\item{pheno_only}{Boolean. If TRUE, only return measurements for
those nodes listed in \code{phenotypes}.}
}
\description{
Process the results of combination perturbation screen, assuming
two perturbations
}
\seealso{
Other combination_therapy: 
\code{\link{check_drug_conflicts}()},
\code{\link{check_drug_nodes}()},
\code{\link{check_drugs_in_range}()},
\code{\link{combo_parallel}()},
\code{\link{get_drugs_commands}()},
\code{\link{make_bkg_commands_combo}()},
\code{\link{make_command_args}()},
\code{\link{make_pair_muts}()},
\code{\link{make_single_muts}()},
\code{\link{run_all_backgrounds}()}
}
\concept{combination_therapy}
