% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_combo_results.r
\name{split_combo_results}
\alias{split_combo_results}
\title{split_combo_results}
\usage{
split_combo_results(
  results_prefix,
  project_path,
  out_dir,
  netw_file_path,
  drug_path = NA
)
}
\arguments{
\item{results_prefix}{prefix to results directory}

\item{project_path}{project path for git SHA log, point to git
repo of the network and specification being tested}

\item{out_dir}{output directory}

\item{netw_file_path}{path to network JSON file}

\item{drug_path}{path to list of drug perturbations to be applied
in addition to node perturbations (CSV)}

\item{calc_survival}{calculate survival meta-phenotype TRUE/FALSE
where survival is \code{survPhenoA - survPhenoB}}

\item{survPhenoA}{Postive phenotype for survival
e.g. Proliferation}

\item{survPhenoB}{Negative phenotype for survival e.g. Apoptosis}
}
\value{
write the results, with survival if enabled, for node,
druggable (only nodes for which there are drugs) and drug
perturbations
}
\description{
Split combination results into node, druggable and drug
perturbations and optionally calculate the difference between one
positive and one negative characteristic e.g. survival
}
