% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_missing_nodes_expected.r
\name{stop_missing_nodes_expected}
\alias{stop_missing_nodes_expected}
\title{stop_missing_nodes_expected}
\usage{
stop_missing_nodes_expected(
  spec,
  missing_nodes_expected_overide = FALSE,
  netw_variables,
  log_file
)
}
\arguments{
\item{spec}{specification dataframe generated by \code{\link{import_spec}}}

\item{missing_nodes_expected_overide}{allows overide of this check}

\item{netw_variables}{network variables dataframe generated by
\code{\link{get_netw_variables}}}

\item{log_file}{name of log file}
}
\value{
stop if nodes expected/measured in specification are missing from network
}
\description{
Stop if there are nodes which are measured/expected in the specification
that are not in the network
}
