% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_missing_nodes_perturbed.r
\name{stop_missing_nodes_perturbed}
\alias{stop_missing_nodes_perturbed}
\title{stop_missing_nodes_perturbed}
\usage{
stop_missing_nodes_perturbed(
  spec,
  missing_nodes_perturbed_overide = FALSE,
  netw_variables,
  log_file
)
}
\arguments{
\item{spec}{specification dataframe generated by \code{\link{import_spec}}}

\item{missing_nodes_perturbed_overide}{allows overide of this check}

\item{netw_variables}{network variables dataframe generated by
\code{\link{get_netw_variables}}}

\item{log_file}{name of log file}
}
\value{
stop if nodes perturbed in specification are missing from network
}
\description{
Stop if there are nodes which are perturbed in the specification
that are not in the network
}
