% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_no_inputs.r
\name{stop_no_inputs}
\alias{stop_no_inputs}
\title{stop_no_inputs}
\usage{
stop_no_inputs(spec, log_file, group_vars)
}
\arguments{
\item{spec}{specification dataframe generated by \code{\link{import_spec}}}

\item{log_file}{log_file name}

\item{group_vars}{variables used to group experiments}
}
\value{
stop if there is an experiment with no inputs
}
\description{
Stop if there is an experiment that does not perturb the network
at all as usually this indicates a user error in the formating of
the specification
}
