% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_autopert_config.r
\name{validate_autopert_config}
\alias{validate_autopert_config}
\title{validate_autopert_config}
\usage{
validate_autopert_config(config_path)
}
\arguments{
\item{config_path}{path to JSON configuration file}
}
\value{
list with validated configuration parameters
}
\description{
Validate autopert configuration from JSON file
}
