% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NANSEN-package.R
\docType{package}
\name{NANSEN-package}
\alias{NANSEN}
\alias{NANSEN-package}
\title{NANSEN: NANSEN (Network Analysis aNd ScrEeNing)}
\description{
A package for verification and analysis of biological gene regulatory network models built with Bio Model Analyzer.
}
\author{
\strong{Maintainer}: Matthew Clarke \email{matthew.clarke@ucl.ac.uk} (\href{https://orcid.org/0000-0002-2582-9689}{ORCID})

Authors:
\itemize{
  \item Pedro Victori \email{pedro.victori@ucl.ac.uk} (\href{https://orcid.org/0000-0001-6321-5584}{ORCID})
}

}
\keyword{internal}
