% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_funcs.r
\name{run_all_backgrounds}
\alias{run_all_backgrounds}
\title{Given a set of backgrounds and a set of perturbations, annotated
with columns of BMA command arguments and corresponding filenames,
passes these to BioCheckConsole to generate results with VMCAI}
\usage{
run_all_backgrounds(
  background_commands,
  netw_file_path,
  bma_path,
  perts,
  combo_type,
  results_dir,
  log_file,
  precedence = "perturbation",
  short_filenames = FALSE,
  file_hashtable_dir = NULL
)
}
\arguments{
\item{background_commands}{data frame of backgrounds generated by
\code{\link{make_bkg_commands_combo}}}

\item{netw_file_path}{path to network JSON file}

\item{bma_path}{path to BMA command line installation.}

\item{perts}{data frame of pertubations e.g. generated by
\code{\link{make_single_muts}} or \code{\link{make_pair_muts}}}

\item{combo_type}{e.g. "single" or "double":used to label output
directories}

\item{results_dir}{directory to store results files}

\item{log_file}{name of log file used to track
results/warnings/errors with futile.logger}

\item{precedence}{should perturbation or background mutation take
precedence if there are conflicting effects on the same node,
defaults to perturbation}

\item{short_filenames}{logical. If TRUE, use MD5 hashes for long filenames
to avoid Windows path length issues. Defaults to FALSE.}

\item{file_hashtable_dir}{directory path where filename hashtables are stored
when short_filenames=TRUE. Defaults to NULL.}
}
\value{
Results are written as json files by BMA in directories
created with the name of the background and the
combo_type. Individual files are named by combining the file
part of the background and the perturbation as generated by
\code{\link{make_bkg_commands_combo}} and
\code{\link{make_single_muts}} or \code{\link{make_pair_muts}}
respectively.
}
\description{
Given a set of backgrounds and a set of perturbations, annotated
with columns of BMA command arguments and corresponding filenames,
passes these to BioCheckConsole to generate results with VMCAI
}
\seealso{
Other combination_therapy: 
\code{\link{check_drug_conflicts}()},
\code{\link{check_drug_nodes}()},
\code{\link{check_drugs_in_range}()},
\code{\link{combo_parallel}()},
\code{\link{get_drugs_commands}()},
\code{\link{make_bkg_commands_combo}()},
\code{\link{make_command_args}()},
\code{\link{make_pair_muts}()},
\code{\link{make_single_muts}()},
\code{\link{process_results}()}
}
\concept{combination_therapy}
